kmm: rename bmc test package directory to 1bmc
